+++ server/src/config.ts
@@ -3,7 +3,6 @@ export interface VLSFormatConfig {
     [lang: string]: string;
   };
   defaultFormatterOptions: {
-    prettier: any;
     [lang: string]: any;
   };
   scriptInitialIndent: boolean;